requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cryptography==37.0.4
dataclasses==0.8
dnspython==2.2.1
docopt==0.6.2
Flask==2.0.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pipreqs==0.4.11
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
typing_extensions==4.1.1
urllib3==1.26.11
Werkzeug==2.0.3
yarg==0.1.9
zipp==3.6.0
